# api/open-api/swagger.yaml
openapi: 3.1.0
info:
  title: Swagger - OpenAPI 3.1
  description: API documentation for the final project.
  # termsOfService: http://swagger.io/terms/
  # contact:
  #   email: apiteam@swagger.io
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
# externalDocs:
#   description: Find out more about Swagger
#   url: http://swagger.io
# servers:
#   - url: https://petstore3.swagger.io/api/v3
...
paths:
  /api/users/register:
    post:
      tags:
        - users
      summary: Create user
      description: Create a new user
      operationId: createUser
      requestBody:
        description: Create user object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
              required:
                - username
                - email
                - password
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: User already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/users/login:
    post:
      tags:
        - users
      summary: Logs user into the system
      description: Creates a new token for the user
      operationId: loginUser
      parameters:
        - name: username
          in: query
          description: The user name for login
          required: false
          schema:
            type: string
        - name: password
          in: query
          description: The password for login in clear text
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Succesful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
          headers:
            token:
              description: The token to be used for authentication
              schema:
                type: string
          #   X-Rate-Limit:
          #     description: Calls per hour allowed by the user
          #     schema:
          #       type: integer
          #       format: int32
          #   X-Expires-After:
          #     description: Date in UTC when token expires
          #     schema:
          #       type: string
          #       format: date-time
        '400':
          description: Invalid username/password supplied
        '404':
          description: User not found
...
  /api/listings:
    post:
      tags:
        - listings
      summary: Create a new listing
      description: This can only be done by the logged in user
      operationId: createListing
      requestBody:
        description: Create listing object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Listing'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Listing'
        '400':
          description: Listing already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
    get:
      tags:
        - listings
      summary: List all listings
      description: This can only be done by the logged in user
      operationId: listListings
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  anyOf:
                    - $ref: '#/components/schemas/Listing'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/listings/{listingId}:
    get:
      tags:
        - listings
      summary: Retrieve details of a specific listing
      description: This can only be done by the logged in user
      operationId: getListingById
      parameters:
        - name: listingId
          in: path
          description: The ID of the listing to retrieve
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Listing'
        '400':
          description: Bad request
        '404':
          description: Listing not found
    delete:
      tags:
        - listings
      summary: Delete listig
      description: This can only be done by the logged in user
      operationId: deleteListing
      parameters:
        - name: listingId
          in: path
          description: The name of the listing that needs to be deleted
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Succesful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Bad request
        '404':
          description: Listing not found
  /api/listings/{listingId}/enable:
    post:
      tags:
        - listings
      summary: Enable listing
      description: This can only be done by the logged in user
      operationId: enableListing
      parameters:
        - name: listingId
          in: path
          description: The ID of the listing to enable
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Bad request
        '404':
          description: Listing not found
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/listings/{listingId}/disable:
    post:
      tags:
        - listings
      summary: Disable listing
      description: This can only be done by the logged in user
      operationId: disableListing
      parameters:
        - name: listingId
          in: path
          description: The ID of the listing to disable
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Bad request
        '404':
          description: Listing not found
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/devices:
    post:
      tags:
        - devices
      summary: Create a new device
      description: This can only be done by the logged in user
      operationId: createDevice
      requestBody:
        description: Create device object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Device'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Device'
        '400':
          description: Device already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
    get:
      tags:
        - devices
      summary: List all devices
      description: This can only be done by the logged in user
      operationId: listDevices
      query:
        - name: listingId
          in: query
          description: The ID of the listing to retrieve devices from
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  anyOf:
                    - $ref: '#/components/schemas/Device'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/devices/{deviceId}:
    patch:
      tags:
        - devices
      summary: Update details of a specific device
      operationId: updateDevice
      parameters:
        - name: deviceId
          in: path
          description: The ID of the device to modify
          required: true
          schema:
            type: string
      requestBody:
        description: Update an existent device object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Device'
      responses:
        '200':
          description: Succesful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Bad request
        '404':
          description: Device not found
...
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
          examples: [10]
        username:
          type: string
          examples: [theUser]
        firstName:
          type: string
          examples: [John]
        lastName:
          type: string
          examples: [James]
        email:
          type: string
          examples: [john@email.com]
        password:
          type: string
          examples: ['12345']
        phone:
          type: string
          examples: ['12345']
        userStatus:
          type: integer
          description: User Status
          format: int32
          examples: [1]
...
    Listing:
      type: object
      properties:
        listingId:
          type: string
          examples: [1234567890]
        provider:
          type: string
          examples: [Airbnb, VRBO, HomeAway]
        title:
          type: string
          examples: [theListing]
        description:
          type: string
          examples: [This is a listing]
        location:
          type: string
          examples: [New York]
        userId:
          type: string
          examples: [1234567890]
        enabled:
          type: boolean
          examples: [true]
    Device:
      type: object
      properties:
        deviceName:
          type: string
          examples: [1234567890]
        deviceType:
          type: string
          examples: [Phone, Tablet, Laptop]
        deviceToken:
          type: string
          examples: [1234567890]
        listingId:
          type: string
          examples: [1234567890]
        deviceId:
          type: string
          examples: [1234567890]
...
    ApiResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
  securitySchemes:
    api_key:
      type: apiKey
      name: api_key
      in: header